+++ plugin/src/main/java/io/snyk/languageserver/protocolextension/SnykExtendedLanguageClient.java
@@ -7,30 +7,31 @@ import java.util.HashSet;
 import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 
-import org.eclipse.jface.dialogs.MessageDialog;
+import org.eclipse.core.resources.IProject;
+import org.eclipse.jdt.internal.core.JavaProject;
+import org.eclipse.jface.viewers.IStructuredSelection;
+import org.eclipse.jface.wizard.WizardDialog;
 import org.eclipse.lsp4e.LSPEclipseUtils;
 import org.eclipse.lsp4e.LanguageClientImpl;
-import org.eclipse.lsp4e.LanguageServerWrapper;
 import org.eclipse.lsp4e.ServerMessageHandler;
 import org.eclipse.lsp4j.ExecuteCommandParams;
 import org.eclipse.lsp4j.Location;
-import org.eclipse.lsp4j.MessageActionItem;
 import org.eclipse.lsp4j.MessageParams;
 import org.eclipse.lsp4j.MessageType;
 import org.eclipse.lsp4j.ProgressParams;
 import org.eclipse.lsp4j.ShowDocumentParams;
 import org.eclipse.lsp4j.ShowDocumentResult;
-import org.eclipse.lsp4j.ShowMessageRequestParams;
 import org.eclipse.lsp4j.WorkDoneProgressCreateParams;
 import org.eclipse.lsp4j.jsonrpc.services.JsonNotification;
-import org.eclipse.swt.SWT;
-import org.eclipse.swt.widgets.Display;
-import org.eclipse.swt.widgets.Shell;
+import org.eclipse.ui.ISelectionService;
+import org.eclipse.ui.IWorkbenchWindow;
 import org.eclipse.ui.PlatformUI;
 
 import io.snyk.eclipse.plugin.SnykStartup;
 import io.snyk.eclipse.plugin.properties.preferences.Preferences;
 import io.snyk.eclipse.plugin.utils.SnykLogger;
+import io.snyk.eclipse.plugin.views.SnykView;
+import io.snyk.eclipse.plugin.wizards.SnykWizard;
 import io.snyk.languageserver.protocolextension.messageObjects.HasAuthenticatedParam;
 import io.snyk.languageserver.protocolextension.messageObjects.SnykIsAvailableCliParams;
 import io.snyk.languageserver.protocolextension.messageObjects.SnykTrustedFoldersParams;
@@ -50,8 +51,43 @@ public class SnykExtendedLanguageClient extends LanguageClientImpl {
     return instance; // we leave instantiation to LSP4e, no lazy construction here
   }
 
-  public void triggerScan() {
-    ExecuteCommandParams params = new ExecuteCommandParams("snyk.workspace.scan", new ArrayList<>());
+  public void triggerScan(IWorkbenchWindow window) {
+    if (Preferences.getInstance().getAuthToken().isBlank()) {
+      runSnykWizard();
+    } else {
+      ExecuteCommandParams params = new ExecuteCommandParams("snyk.workspace.scan", new ArrayList<>());
+      try {
+        getLanguageServer().getWorkspaceService().executeCommand(params);
+
+        if (window == null) {
+          window = PlatformUI.getWorkbench().getActiveWorkbenchWindow();
+        }
+        if (window == null) {
+          return;
+        }
+        ISelectionService service = window.getSelectionService();
+        IStructuredSelection structured = (IStructuredSelection) service.getSelection();
+
+        Object firstElement = structured.getFirstElement();
+
+        if (firstElement instanceof IProject) {
+          IProject project = (IProject) firstElement;
+          runForProject(project.getName());
+        }
+
+        if (firstElement instanceof JavaProject) {
+          JavaProject javaproject = (JavaProject) firstElement;
+          runForProject(javaproject.getProject().getName());
+        }
+      } catch (Exception e) {
+        SnykLogger.logError(e);
+      }
+
+    }
+  }
+
+  public void triggerAuthentication() {
+    ExecuteCommandParams params = new ExecuteCommandParams("snyk.login", new ArrayList<>());
     try {
       getLanguageServer().getWorkspaceService().executeCommand(params);
     } catch (Exception e) {
@@ -62,8 +98,11 @@ public class SnykExtendedLanguageClient extends LanguageClientImpl {
   @JsonNotification(value = "$/snyk.hasAuthenticated")
   public void hasAuthenticated(HasAuthenticatedParam param) {
     Preferences.getInstance().store(Preferences.AUTH_TOKEN_KEY, param.getToken());
-    showAuthenticatedMessage();
-    enableSnykViewRunActions();
+    triggerScan(null);
+    if (!param.getToken().isBlank()) {
+      showAuthenticatedMessage();
+      enableSnykViewRunActions();
+    }
   }
 
   @JsonNotification(value = "$/snyk.isAvailableCli")
@@ -97,6 +136,15 @@ public class SnykExtendedLanguageClient extends LanguageClientImpl {
     progressMgr.updateProgress(params);
   }
 
+  private void runSnykWizard() {
+    SnykWizard wizard = new SnykWizard();
+
+    WizardDialog dialog = new WizardDialog(PlatformUI.getWorkbench().getDisplay().getActiveShell(), wizard);
+
+    dialog.setBlockOnOpen(true);
+    dialog.open();
+  }
+
   private void enableSnykViewRunActions() {
     PlatformUI.getWorkbench().getDisplay().asyncExec(() -> {
       var snykView = SnykStartup.getSnykView();
@@ -110,6 +158,13 @@ public class SnykExtendedLanguageClient extends LanguageClientImpl {
     messageParams.setType(MessageType.Info);
     messageParams.setMessage("The authentication token has been stored in Snyk Preferences.");
     ServerMessageHandler.showMessage("Authentication with Snyk successful", messageParams);
+  }
+
+  private void runForProject(String projectName) {
+    SnykView snykView = SnykStartup.getSnykView();
+    if (snykView != null) {
+      snykView.testProject(projectName);
+    }
   }
 
   // TODO: remove once LSP4e supports `showDocument` in its next release (it's
@@ -128,7 +183,7 @@ public class SnykExtendedLanguageClient extends LanguageClientImpl {
       return new ShowDocumentResult(true);
     });
   }
-  
-  
-  
+
+
+
 }