+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Snyk Changelog
 
+
 ## [2.0.0] - Unreleased
 ### Changes
 - add workspace trust feature
+- adds configuration wizard for custom endpoints
+
+## [2.0.0] - v20221007.135736
+### Fixed
+- malformed proxy URL
 
 ## [2.0.0] - v20220927.182222
 ### Changes